[feat] Modularização dos componentes
